requirements.txt
python-telegram-bot[job-queue]==20.7
aiohttp==3.9.1
APScheduler==3.10.4
pytz==2024.1